nexus_agent_box/pyproject.toml
[project]
name = "nexus-agent-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b-code-interpreter>=2.6.0",
    "fastapi>=0.136.1",
    "modelscope>=1.36.1",
    "oss2>=2.19.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
